fixed rf
